~ fix 12 bugs across market engine, plugins, and utilities
Fixes: switch fallthrough in VirtualOrderBookCache, PID div-by-zero,
ServerMarket candle state on load, AsyncMarket missing settings param,
volume distribution init spike, Orderbook undercounting at market price,
ServerMarketManager discarding markets on partial failure, DataManager
save lock leak, PriceHistoryData integer bounds, DynamicIndexedArray
source offset, and null guards in TargetPriceBot and VolumeDistribution.
Adds 7 regression tests (264 total).

+++ common/src/main/java/net/kroia/stockmarket/util/PriceHistoryData.java
@@ -273,7 +273,7 @@ public class PriceHistoryData
     {
         if(candles.isEmpty())
             return 0;
-        long min = Integer.MAX_VALUE;
+        long min = Long.MAX_VALUE;
         if(startIndex > endIndex)
         {
             int tmp = endIndex;
@@ -292,7 +292,7 @@ public class PriceHistoryData
     {
         if(candles.isEmpty())
             return 0;
-        long max = Integer.MIN_VALUE;
+        long max = Long.MIN_VALUE;
         if(startIndex > endIndex)
         {
             int tmp = endIndex;
